Separate out ReporterStats helper classes
